% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbfs.R
\name{get_gbfs_cities}
\alias{get_gbfs_cities}
\title{Get table of all cities releasing GBFS feeds}
\source{
North American Bikeshare Association, General Bikeshare Feed Specification
 \url{https://raw.githubusercontent.com/MobilityData/gbfs/master/systems.csv}
}
\usage{
get_gbfs_cities()
}
\value{
A \code{data.frame} of all cities issuing GBFS feeds. The `Auto-Discovery URL`
column supplies the relevant .json feeds, while the entries in the `URL` column 
take the user to the public-facing webpage of the programs.
}
\description{
Get table of all cities releasing GBFS feeds
}
