% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dynamic_feeds.R
\name{get_station_status}
\alias{get_station_status}
\title{Save the station_status file as a .rds file.}
\usage{
get_station_status(url, filepath = "data/station_status.rds",
  cumulative = FALSE)
}
\arguments{
\item{url}{A link to an active station_status .json feed.}

\item{filepath}{A connection or path to save the .rds to-- defaults to 
"data/station_status.rds"}
}
\value{
A .rds object generated from the current specified feed.
}
\description{
Save the station_status file as a .rds file.
}
\examples{
get_station_status(url = "http://biketownpdx.socialbicycles.com/opendata/station_status.json")
}
