% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_static_feeds.R
\name{get_system_alerts}
\alias{get_system_alerts}
\title{Save the system_alerts file as a .rds file.}
\usage{
get_system_alerts(url, filepath = "data/system_alerts.rds")
}
\arguments{
\item{url}{A link to an active system_alerts .json feed.}

\item{filepath}{A connection or path to save the .rds to-- defaults to 
"data/system_alerts.rds"}
}
\value{
A .rds object generated from the current specified feed.
}
\description{
Save the system_alerts file as a .rds file.
}
\examples{
get_system_alerts(url = "http://biketownpdx.socialbicycles.com/opendata/system_alerts.json")
}
