% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_static_feeds.R
\name{get_system_calendar}
\alias{get_system_calendar}
\title{Save the system_calendar file as a .rds file.}
\usage{
get_system_calendar(url, filepath = "data/system_calendar.rds")
}
\arguments{
\item{url}{A link to an active system_calendar .json feed.}

\item{filepath}{A connection or path to save the .rds to-- defaults to 
"data/system_calendar.rds"}
}
\value{
A .rds object generated from the current specified feed.
}
\description{
Save the system_calendar file as a .rds file.
}
\examples{
get_system_calendar(url = "http://biketownpdx.socialbicycles.com/opendata/system_calendar.json")
}
