% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_static_feeds.R
\name{get_system_hours}
\alias{get_system_hours}
\title{Save the system_hours file as a .rds file.}
\usage{
get_system_hours(url, filepath = "data/system_hours.rds")
}
\arguments{
\item{url}{A link to an active system_hours .json feed.}

\item{filepath}{A connection or path to save the .rds to-- defaults to 
"data/system_hours.rds"}
}
\value{
A .rds object generated from the current specified feed.
}
\description{
Save the system_hours file as a .rds file.
}
\examples{
get_system_hours(url = "http://biketownpdx.socialbicycles.com/opendata/system_hours.json")
}
