% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_static_feeds.R
\name{get_system_information}
\alias{get_system_information}
\title{Save the system_information file as a .rds file.}
\usage{
get_system_information(url, filepath = "data/system_information.rds")
}
\arguments{
\item{url}{A link to an active system_information .json feed.}

\item{filepath}{A connection or path to save the .rds to-- defaults to 
"data/system_information.rds"}
}
\value{
A .rds object generated from the current specified feed.
}
\description{
Save the system_information file as a .rds file.
}
\examples{
get_system_information(url = "http://biketownpdx.socialbicycles.com/opendata/system_information.json")
}
