% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_static_feeds.R
\name{get_system_pricing_plans}
\alias{get_system_pricing_plans}
\title{Save the system_pricing_plans file as a .rds file.}
\usage{
get_system_pricing_plans(url, filepath = "data/system_pricing_plans.rds")
}
\arguments{
\item{url}{A link to an active system_pricing_plans .json feed.}

\item{filepath}{A connection or path to save the .rds to-- defaults to 
"data/system_pricing_plans.rds"}
}
\value{
A .rds object generated from the current specified feed.
}
\description{
Save the system_pricing_plans file as a .rds file.
}
\examples{
get_system_pricing_plans(url = "http://biketownpdx.socialbicycles.com/opendata/system_pricing_plans.json")
}
