% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_static_feeds.R
\name{get_system_regions}
\alias{get_system_regions}
\title{Save the system_regions file as a .rds file.}
\usage{
get_system_regions(url, filepath = "data/system_regions.rds")
}
\arguments{
\item{url}{A link to an active system_regions .json feed.}

\item{filepath}{A connection or path to save the .rds to-- defaults to 
"data/system_regions.rds"}
}
\value{
A .rds object generated from the current specified feed.
}
\description{
Save the system_regions file as a .rds file.
}
\examples{
get_system_regions(url = "http://biketownpdx.socialbicycles.com/opendata/system_regions.json")
}
