% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbfs_description.R
\docType{package}
\name{gbfs}
\alias{gbfs}
\alias{gbfs-package}
\title{Package: gbfs}
\description{
General Bikeshare Feed Specification
}
\details{
The `gbfs` package supplies a set of functions to interface with General
Bikeshare Feed Specification .json feeds in R, allowing users to save
and accumulate tidy .rds datasets for specified cities/bikeshare programs.
}
\section{gbfs Functions}{

get_gbfs(), get_free_bike_status(), get_station_information(),
get_station_status(), get_system_alerts(), get_system_calendar(),
get_system_hours(), get_system_information(),
get_system_pricing_plans(), get_system_regions()
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ds-civic-data/gbfs}
  \item Report bugs at \url{https://github.com/ds-civic-data/gbfs/issues}
}

}
\author{
\strong{Maintainer}: Simon P. Couch \email{simonpatrickcouch@gmail.com}

Authors:
\itemize{
  \item Kaelyn Rosenberg \email{kaerosenberg@gmail.com}
}

Other contributors:
\itemize{
  \item Mark Padgham \email{mark.padgham@email.com} [contributor]
}

}
