% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dynamic_feeds.R
\name{get_free_bike_status}
\alias{get_free_bike_status}
\title{Save the free_bike_status file as a .rds file.}
\usage{
get_free_bike_status(url, filepath = "data/free_bike_status.rds")
}
\arguments{
\item{url}{A link to an active free_bike_status .json feed.}

\item{filepath}{A connection or path to save the .rds to-- defaults to 
"data/free_bike_status.rds"}
}
\value{
A .rds object generated from the current specified feed.
}
\description{
Save the free_bike_status file as a .rds file.
}
\examples{
get_free_bike_status(url = "http://biketownpdx.socialbicycles.com/opendata/free_bike_status.json")
}
