% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbfs.R
\name{get_gbfs}
\alias{get_gbfs}
\title{Save gbfs feeds.}
\usage{
get_gbfs(city, feeds = "all", directory)
}
\arguments{
\item{city}{A character string that can be matched to a city or a url to an 
active gbfs.json feed. See \code{get_gbfs_cities} for a current list of available cities.}

\item{feeds}{A character string specifying which feeds should be saved. Options are
\code{"all"}, \code{"static"}, and \code{"dynamic"}.}

\item{directory}{The name of an existing folder or folder to be created, where the feeds
will be saved.}
}
\value{
A folder containing the specified feeds saved as .rds objects.
}
\description{
\code{get_gbfs} checks for the existence of General Bikeshare Specification
feeds for a given city and saves the feeds as .rds objects in a directory that
can be specified by the user. Metadata for each dataset can be found at:
`https://github.com/NABSA/gbfs/blob/master/gbfs.md`
}
\examples{
# to grab all of the feeds released by portland, oregon's
# bikeshare program "biketown",
\donttest{get_gbfs(city = "portland", directory = tempdir())}

# note that, usually, we'd supply a character string 
# (like "pdx", maybe,) to the directory argument 
# instead of `tempdir()`. 

# if we're having trouble specifying the correct feed,
# we can also supply the actual URL to the feed
\donttest{get_gbfs(city = "http://biketownpdx.socialbicycles.com/opendata/gbfs.json", 
         directory = tempdir())}
                   
# the examples above grab every feed that portland releases.
# if, instead, we just wanted the dynamic feeds,
\donttest{get_gbfs(city = "portland", 
         directory = tempdir(),
         feeds = "dynamic")}
}
