% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbfs.R
\name{get_gbfs}
\alias{get_gbfs}
\title{Save gbfs feeds to a user-specified directory.}
\usage{
get_gbfs(city, feeds = "all", directory = "gbfs_data")
}
\arguments{
\item{city}{A character string or a url to an active gbfs.json feed}

\item{directory}{A link to an active station_information .json feed.}

\item{filepath}{A folder to save the files to--defaults to "gbfs_data"}
}
\value{
A .rds object generated from the current specified feed.
}
\description{
Save gbfs feeds to a user-specified directory.
}
\examples{
get_gbfs(city = "topeka")
get_gbfs(city = "http://biketownpdx.socialbicycles.com/opendata/gbfs.json")
get_gbfs(city = "http://biketownpdx.socialbicycles.com/opendata/gbfs.json")
get_gbfs("https://gbfs.bcycle.com/bcycle_greenbikeslc/gbfs.json", directory = "slcbikes")
}
