% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbfs.R
\name{get_gbfs_cities}
\alias{get_gbfs_cities}
\title{Get table of all cities releasing GBFS feeds}
\usage{
get_gbfs_cities()
}
\value{
A \code{data.frame} of all cities issuing GBFS feeds
}
\description{
Get table of all cities releasing GBFS feeds
}
