% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_static_feeds.R
\name{get_station_information}
\alias{get_station_information}
\title{Save the station_information file as a .rds file.}
\usage{
get_station_information(url, filepath = "data/station_information.rds")
}
\arguments{
\item{url}{A link to an active station_information .json feed.}

\item{filepath}{A connection or path to save the .rds to-- defaults to 
"data/station_information.rds"}
}
\value{
A .rds object generated from the current specified feed.
}
\description{
Save the station_information file as a .rds file.
}
\examples{
get_station_information(url = "http://biketownpdx.socialbicycles.com/opendata/station_information.json")
}
