% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_static_feeds.R
\name{get_system_calendar}
\alias{get_system_calendar}
\title{Grab the system_calendar feed.}
\usage{
get_system_calendar(
  city,
  directory = NULL,
  file = "system_calendar.rds",
  output = "save"
)
}
\arguments{
\item{city}{A character string that can be matched to a city or a url to an 
active gbfs .json feed. See \code{get_gbfs_cities} for a current list of available cities.}

\item{directory}{The name of an existing folder or folder to be created, where the feed will
will be saved. This argument is ignored if \code{output = "return"}.}

\item{file}{The name of the file to be saved (if \code{output} is set to \code{"save"} 
or \code{"both"}), as a character string. Must end in \code{".rds"}.}

\item{output}{The type of output method. If `output = "save"`, the object will be saved as
an .rds object at the given path. If `output = "return"`, the output will be returned
as a dataframe object. Setting `output = "both"` will do both.}
}
\value{
The output of this function depends on the argument to `output`: 
Either a saved .rds object generated from the 
current system_calendar feed, a dataframe object, or both
}
\description{
\code{get_system_calendar} grabs and tidies the system_calendar feed for a given city. See
`https://github.com/NABSA/gbfs/blob/master/gbfs.md` to see metadata for this dataset.
}
\examples{
# we can grab the system calendar feed for portland, 
# oregon in one of several ways! first, supply the `city` 
# argument as a URL, and save to file by leaving output 
# set to it's default. usually, we would supply a character 
# string (like "pdx", maybe,) for the `directory` argument 
# instead of `tempdir`.
\donttest{get_system_calendar(city = "http://biketownpdx.socialbicycles.com/opendata/system_calendar.json",  
                    directory = tempdir())}
                    
# or, instead, just supply the name of 
# the city as a string. 
\donttest{get_system_calendar(city = "portland",
                    directory = tempdir())}
                    
# instead of saving the output as a file, we can 
# just return the output as a dataframe
\donttest{get_system_calendar(city = "portland",  
                    output = "return")}
}
